+++ src/moo-color.ts
@@ -24,12 +24,17 @@ export class MooColor extends ColorFormatter implements ColorModifiable<MooColor
 
   /**
    * Creates an instance of MooColor.
-   * @param {*} [color] color string. e.g. '#ff0000' 'rgba(255, 0, 0, .5)' 'hsl(120, 50%, 100%)'
+   * @param {(string|Color)} [color] color value. e.g. '#ff0000' 'rgba(255, 0, 0, .5)' 'hsl(120, 50%, 100%)'
+   * @memberof MooColor
    */
-  constructor(color?: any) {
+  constructor(color?: string|Color) {
     super();
-    color = color ? color : '#000';
-    this.setColorByParser(color);
+    if (typeof color === 'object' && color !== null) {
+      this.setColor(color as Color);
+    } else if (typeof color === 'string' || typeof color === 'undefined') {
+      color = color ? color : '#000';
+      this.setColorByParser(color as string);
+    }
   }
 
   setColorByParser(str: string): this {