+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Now available give color argument as `ColorData` in MooColor constructor. #7
+
 ## [0.1.3] - 2018-03-30
 
 ### Added